Clean up ActionButton and Button
